update function to be long running
