+++ samples/ingestion/ingestion-client/FetchTranscription/host.json
@@ -1,4 +1,4 @@
 {
   "version": "2.0",
-  "functionTimeout": "00:10:00"
+  "functionTimeout": "00:30:00"
 }
